delete now only works for logged in user owning that account

+++ api/queries/check_ins.py
@@ -287,16 +287,22 @@ class Check_InQueries:
         except Exception:
             return Error(message="Could not update check-in!")
 
-    def delete(self, check_in_id: int) -> bool:
+    def delete(self, check_in_id: int, account_data: dict) -> Union[bool, Error]:
+        check_in = self.get_one_check_in(
+            check_in_id=check_in_id,
+            account_data=account_data
+        )
+        if isinstance(check_in, Error):
+            return check_in
         try:
             with pool.connection() as conn:
                 with conn.cursor() as db:
                     db.execute(
                         """
                         DELETE FROM check_ins
-                        WHERE check_in_id = %s;
+                        WHERE check_in_id = %s AND account = %s;
                         """,
-                        [check_in_id],
+                        [check_in_id, account_data['id']],
                     )
                     return True
         except Exception:
